enhance: Make types adaptive to older versions
